+++ core/src/main/java/io/snabble/sdk/ProductDatabase.java
@@ -23,6 +23,7 @@ import java.util.Date;
 import java.util.List;
 import java.util.Scanner;
 import java.util.concurrent.CopyOnWriteArrayList;
+import java.util.stream.Stream;
 
 import io.snabble.sdk.codes.ScannedCode;
 import io.snabble.sdk.codes.templates.CodeTemplate;
@@ -116,7 +117,7 @@ public class ProductDatabase {
                 schemaVersionMajor = Integer.parseInt(getMetaData(METADATA_KEY_SCHEMA_VERSION_MAJOR));
                 schemaVersionMinor = Integer.parseInt(getMetaData(METADATA_KEY_SCHEMA_VERSION_MINOR));
 
-                if (schemaVersionMajor == 1 && schemaVersionMinor < 24) {
+                if (schemaVersionMajor == 1 && schemaVersionMinor < 25) {
                     Logger.d("Database has incompatible schema, deleting local database");
                     delete();
                     return false;
@@ -769,6 +770,7 @@ public class ProductDatabase {
         int[] codeSpecifiedQuantities = null;
         boolean[] codeIsPrimaryCode = null;
         String[] templates = null;
+        String[] transmissionTemplates = null;
 
         String sku = anyToString(cursor, 0);
 
@@ -867,9 +869,14 @@ public class ProductDatabase {
                     }
                 }
             }
+
+            String transmissionTemplatesStr = cursor.getString(19);
+            if (transmissionTemplatesStr != null) {
+                transmissionTemplates = transmissionTemplatesStr.split(",", -1);
+            }
         }
 
-        String scanMessage = cursor.getString(19);
+        String scanMessage = cursor.getString(20);
         if (scanMessage != null) {
             builder.setScanMessage(scanMessage);
         }
@@ -881,6 +888,7 @@ public class ProductDatabase {
                 String transmissionCode = null;
                 Unit codeEncodingUnit = null;
                 CodeTemplate template = null;
+                CodeTemplate transmissionTemplate = null;
 
                 if (transmissionCodes != null) {
                     String tc = transmissionCodes[i];
@@ -902,6 +910,15 @@ public class ProductDatabase {
                     }
                 }
 
+                if (transmissionTemplates != null) {
+                    String transmissionTemplateStr = transmissionTemplates[i];
+                    for (CodeTemplate codeTemplate : project.getCodeTemplates()) {
+                        if (codeTemplate.getName().equals(transmissionTemplateStr)) {
+                            transmissionTemplate = codeTemplate;
+                        }
+                    }
+                }
+
                 String primaryTransmissionCode = transmissionCode;
                 if (codeIsPrimaryCode != null) {
                     for (int j = 0; j < codeIsPrimaryCode.length; j++) {
@@ -916,9 +933,11 @@ public class ProductDatabase {
                 }
 
                 String templateName = template != null ? template.getName() : null;
+                String transmissionTemplateName = transmissionTemplate != null ? transmissionTemplate.getName() : null;
                 productCodes[i] = new Product.Code(lookupCode,
                         primaryTransmissionCode,
                         templateName,
+                        transmissionTemplateName,
                         codeEncodingUnit,
                         codeIsPrimaryCode != null && codeIsPrimaryCode[i],
                         codeSpecifiedQuantities != null ? codeSpecifiedQuantities[i] : 0);
@@ -927,7 +946,7 @@ public class ProductDatabase {
             builder.setScannableCodes(productCodes);
         }
 
-        int availability = cursor.getInt(20);
+        int availability = cursor.getInt(21);
         Product.Availability[] availabilities = Product.Availability.values();
         if (availability >= 0 && availability < availabilities.length) {
             builder.setAvailability(availabilities[availability]);
@@ -947,7 +966,7 @@ public class ProductDatabase {
 
         String priceQuery;
 
-        if (schemaVersionMinor >= 22 && shop != null) {
+        if (shop != null) {
             priceQuery = "SELECT listPrice, discountedPrice, customerCardPrice, basePrice FROM prices " +
                     "JOIN shops ON shops.pricingCategory = prices.pricingCategory " +
                     "WHERE shops.id = ? AND sku = ? " +
@@ -1002,14 +1021,14 @@ public class ProductDatabase {
         return in;
     }
 
-    private String productSqlString(String appendSql, boolean distinct) {
+    private String productSqlString(String appendFields, String appendSql, boolean distinct) {
         Shop shop = project.getCheckedInShop();
         String shopId = "0";
         if (shop != null) {
             shopId = shop.getId();
         }
 
-        String sql = "SELECT " + (distinct ? "DISTINCT " : "") +
+        return "SELECT " + (distinct ? "DISTINCT " : "") +
                 "p.sku," +
                 "p.name," +
                 "p.description," +
@@ -1029,16 +1048,20 @@ public class ProductDatabase {
                 ",(SELECT group_concat(ifnull(s.template, \"\")) FROM scannableCodes s WHERE s.sku = p.sku)" +
                 ",(SELECT group_concat(ifnull(sc.isPrimary, '')) FROM scannableCodes sc where sc.sku = p.sku)" +
                 ",(SELECT group_concat(ifnull(sc.specifiedQuantity, '')) FROM scannableCodes sc where sc.sku = p.sku)" +
+                ",(SELECT group_concat(ifnull(sc.transmissionTemplate, '')) FROM scannableCodes sc where sc.sku = p.sku)" +
                 ",p.scanMessage" +
                 ",ifnull((SELECT a.value FROM availabilities a WHERE a.sku = p.sku AND a.shopID = " + shopId + "), " + defaultAvailability + ") as availability" +
+                appendFields +
                 " FROM products p "
                 + appendSql;
-
-        return sql;
     }
 
     private Cursor productQuery(String appendSql, String[] args, boolean distinct, CancellationSignal cancellationSignal) {
-        return rawQuery(productSqlString(appendSql, distinct), args, cancellationSignal);
+        return rawQuery(productSqlString("", appendSql, distinct), args, cancellationSignal);
+    }
+
+    private Cursor productQuery(String appendFields, String appendSql, String[] args, CancellationSignal cancellationSignal) {
+        return rawQuery(productSqlString(appendFields, appendSql, true), args, cancellationSignal);
     }
 
     public boolean isAvailableOffline() {
@@ -1358,6 +1381,36 @@ public class ProductDatabase {
     }
 
     /**
+     * This function needs config value generateSearchIndex set to true
+     * <p>
+     * Returns a {@link Cursor} which can be iterated for items containing the given search
+     * string at the start of a word.
+     * <p>
+     * Matching is normalized, so "appl" finds products containing
+     * "Apple", "apple" and "Super apple", but not "Superapple".
+     *
+     * With this implementation you can JOIN own tables or add conditions and inject orders
+     *
+     * @param cancellationSignal Calls can be cancelled with a {@link CancellationSignal}. Can be null.
+     */
+    public Cursor searchByFoldedName(String attentionalFields, String searchString, CancellationSignal cancellationSignal,
+                                     String attentionalJoins, String attentionalConditions, String... attentionalArgs) {
+        return productQuery(attentionalFields, "JOIN searchByName ns ON ns.sku = p.sku " +
+                        attentionalJoins + " " +
+                        "WHERE ns.foldedName MATCH ? " +
+                        "AND p.weighing != " + Product.Type.PreWeighed.getDatabaseValue() + " " +
+                        "AND p.weighing != " + Product.Type.DepositReturnVoucher.getDatabaseValue() + " " +
+                        "AND p.isDeposit = 0 " +
+                        "AND availability != 2 " +
+                        attentionalConditions + " " +
+                        "LIMIT 100",
+                Stream.of(new String[]{
+                        searchString + "*"
+                }, attentionalArgs).flatMap(Stream::of).toArray(String[]::new),
+                cancellationSignal);
+    }
+
+    /**
      * Returns a {@link Cursor} which can be iterated for items containing the given scannable code.
      * <p>
      * Allows for partial matching. "978" finds products containing "978020137962".
@@ -1387,8 +1440,7 @@ public class ProductDatabase {
         sb.append(" AND p.isDeposit = 0 ");
         sb.append(" AND availability != 2");
 
-        String query = productSqlString(sb.toString(), true) +
-                " LIMIT 100";
+        String query = productSqlString("", sb.toString(), true) + " LIMIT 100";
 
         return rawQuery(query, new String[]{ searchString + "*" }, cancellationSignal);
     }